move surface related utils functions under surface/_utils.py
