+++ nilearn/plotting/html_surface.py
@@ -10,7 +10,7 @@ import numpy as np
 from nilearn._utils import fill_doc
 from nilearn._utils.niimg_conversions import check_niimg_3d
 from nilearn.plotting import cm
-from nilearn.plotting._utils import (
+from nilearn.plotting.surface._utils import (
     check_surface_plotting_inputs,
     sanitize_hemi_for_surface_image,
 )
