+++ nilearn/plotting/surface/surf_plotting.py
@@ -19,11 +19,11 @@ from nilearn import image, surface
 from nilearn._utils import check_niimg_3d, compare_version, fill_doc
 from nilearn._utils.helpers import is_kaleido_installed, is_plotly_installed
 from nilearn._utils.param_validation import check_params
-from nilearn.plotting._utils import (
+from nilearn.plotting.surface._utils import (
     check_surface_plotting_inputs,
-    sanitize_hemi_for_surface_image,
-    save_figure_if_needed,
+    sanitize_hemi_for_surface_image
 )
+from nilearn.plotting._utils import save_figure_if_needed
 from nilearn.plotting.cm import mix_colormaps
 from nilearn.plotting.displays._figures import PlotlySurfaceFigure
 from nilearn.plotting.displays._slicers import get_cbar_ticks
